Turns out I forgot to test if it worked these guys worked properly for the single element.  They did not.  That has now been remedied.

+++ source/ui/VESPERS/XRFViewer.cpp
@@ -343,7 +343,7 @@ void XRFViewer::loadFile()
 			rawData[i] << temp.at(i).toDouble();
 			corrData[i] << temp.at(i+elements).toDouble();
 		}
-		corrSumData << temp.last().toInt();
+		corrSumData << temp.last().toDouble();
 	}
 
 	file.close();
